Fixed a bug where successive runShellCmd() invocations throw exceptions
I was consistently experiencing the following error when querying the
device stats from the watch after the first click of the connect button:

    Error: Incorrect arg0: 0x000000e4 (expected 0x000000e5)

Clicking it a second time correctly retrieved the stats. Also when
querying the list of disabled packages after first disabling one would
always throw the error and there was no way to work around it.

There's no docs or examples of webadb.js that i could find to fix this
problem, but in looking through the source code I saw the .close()
method and it seems to have fixed the problem in my testing.