Added the "Accept ADB connection" modal
Now when the connect button is pressed a modal appears telling the user
to accept the ADB connection on the watch. Once accepted, the modal
disappears but some work will need to be done to allow the user to quit
the modal if the ADB prompt on the watch disappears and they can't
accept it to clear the modal.